Don't rely on UB to handle multiplicity.
Old implementation relied on a fact that ZST are not passed as
arguments, and it appears to work as if argument was omitted from
signature altogether. This behavior is not guaranteed by the compiler,
and may change in the future or work differently on other architectures.

This patch changes the code generator to always generate signatures that
match C code and introduces a macro pthx! to handle calls and
declarations.